@asvetlov 2024-11-21
6% slower in the worst case and no difference in a more realistic scenario sounds good to me.
Thanks!